builder/pyproject.toml
[project]
name = "nixie"
description = "Portable Nix script builder"
authors = [ { "name" = "Karim Vergnes", "email" = "me@thesola.io" } ]
license = { "file" = "LICENSE" }
readme = "README.md"
urls = { "Source" = "https://github.com/nixie-dev/nixie" }
dependencies = [
    "click",
    "rich",
    "click-option-group",
    "python-dotenv",
    "pzp",
    "gitpython"
    ]
dynamic = [ "version" ]

[build-system]
requires = [ "setuptools" ]
build-backend = "setuptools.build_meta"

[project.scripts]
nixie = "nixie.cli:main"

[tool.setuptools.package-data]
"nixie.output" = ["*.in"]
